+++ .gitignore
@@ -24,3 +24,4 @@
 /VstsClientLibrariesSamples/obj/Release
 /VSTSRestApiSamples/obj/Release
 /VSTSRestApiSamples.UnitTests/obj/Release
+/.vs
